refactor components, PascalCase, and clean imports
