+++ frontend/quickcheck/src/components/FetchAllAreas.jsx
@@ -1,0 +1,11 @@
+
+export const AreaMock = {
+  0: { id: 0, name: 'Kredit', category: 'Privat' },
+  1: { id: 1, name: 'Kunden', category: 'Business' },
+  2: { id: 2, name: 'Payments', category: 'Privat' },
+};
+
+
+export const FetchAllAreas = () => {
+  return Object.values(AreaMock);
+};
