+++ frontend/quickcheck/src/components/RemoveMemberButton.jsx
@@ -1,0 +1,44 @@
+import React from 'react';
+import {
+  Button,
+  useDisclosure,
+  IconButton,
+  Popover,
+  PopoverTrigger,
+  PopoverContent,
+  PopoverHeader,
+  PopoverBody,
+} from '@chakra-ui/react';
+import { DeleteIcon} from '@chakra-ui/icons';
+
+//components
+
+
+export default function RemoveButton({ onRemove }) {
+  const { onOpen, onClose, isOpen } = useDisclosure();
+  return (
+    <Popover isOpen={isOpen} onOpen={onOpen} onClose={onClose} isLazy={true} w="wrap">
+      <PopoverTrigger>
+        <IconButton icon={<DeleteIcon />} onClick={onOpen} size="md" variant="wisper" w={16} />
+      </PopoverTrigger>
+      <PopoverContent>
+        <PopoverHeader fontWeight="semibold">Confirm removing this User</PopoverHeader>
+        <PopoverBody>
+          <Button
+            colorScheme="red"
+            mx={1}
+            onClick={(e) => {
+              onRemove();
+              onClose();
+            }}
+          >
+            Remove
+          </Button>
+          <Button mx={1} onClick={onClose}>
+            Cancel
+          </Button>
+        </PopoverBody>
+      </PopoverContent>
+    </Popover>
+  );
+}
