+++ frontend/quickcheck/src/components/RemoveProjectAreaButton.jsx
@@ -1,0 +1,46 @@
+import { React} from 'react';
+import {
+  useDisclosure,
+  Button,
+  Popover,
+  PopoverTrigger,
+  PopoverContent,
+  PopoverHeader,
+  PopoverBody,
+  IconButton,
+} from '@chakra-ui/react';
+import { DeleteIcon } from '@chakra-ui/icons';
+
+//components
+
+
+
+
+export default function RemoveProjectAreaButton({ onRemove }) {
+  const { onOpen, onClose, isOpen } = useDisclosure();
+  return (
+    <Popover isOpen={isOpen} onOpen={onOpen} onClose={onClose} isLazy={true} w="wrap">
+      <PopoverTrigger>
+        <IconButton icon={<DeleteIcon />} onClick={onOpen} variant="wisper" size="md" />
+      </PopoverTrigger>
+      <PopoverContent>
+        <PopoverHeader fontWeight="semibold">Confirm removing this User</PopoverHeader>
+        <PopoverBody>
+          <Button
+            colorScheme="red"
+            mx={1}
+            onClick={(e) => {
+              onRemove();
+              onClose();
+            }}
+          >
+            Remove
+          </Button>
+          <Button mx={1} onClick={onClose}>
+            Cancel
+          </Button>
+        </PopoverBody>
+      </PopoverContent>
+    </Popover>
+  );
+}
